+++ CHANGELOG.md
@@ -3,6 +3,7 @@ CHANGELOG
 
 
 * [FEATURE] #7583 - add badge for hidden cms
+* [FEATURE] #7011 - implement simple search box
 
 4.4.2 (2024-08-02)
 ------------------